Create shopkeeper account #31

+++ backend/src/main/java/fr/univcotedazur/simpletcfs/components/AdminManager.java
@@ -1,7 +1,9 @@
 package fr.univcotedazur.simpletcfs.components;
 
 import fr.univcotedazur.simpletcfs.entities.*;
+import fr.univcotedazur.simpletcfs.exceptions.AlreadyExistingMemberException;
 import fr.univcotedazur.simpletcfs.exceptions.MissingInformationException;
+import fr.univcotedazur.simpletcfs.exceptions.UnderAgeException;
 import fr.univcotedazur.simpletcfs.interfaces.*;
 import fr.univcotedazur.simpletcfs.repositories.AdminAccountRepository;
 import fr.univcotedazur.simpletcfs.repositories.ShopKeeperAccountRepository;
@@ -9,6 +11,7 @@ import fr.univcotedazur.simpletcfs.repositories.ShopRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
@@ -76,20 +79,36 @@ public class AdminManager implements ShopRegistration, ShopkeeperRegistration, A
     }
 
     @Override
-    public ShopKeeperAccount createShopKeeperAccount(Form form, Shop shop) throws MissingInformationException {
+    public ShopKeeperAccount createShopKeeperAccount(Form form, Shop shop) throws MissingInformationException,AlreadyExistingMemberException, UnderAgeException {
         if (form.getName() == null || form.getMail() == null || form.getPassword() == null || form.getBirthDate() == null) {
             throw new MissingInformationException();
         }
-        ShopKeeperAccount shopKeeperAccount = new ShopKeeperAccount(UUID.randomUUID(), form.getName(), form.getMail(), form.getPassword(), form.getBirthDate(), shop);
+        ShopKeeperAccount shopKeeperAccount = findShopkeeperAccountByMail(form.getMail());
+        if(shopKeeperAccount != null ){
+            throw new AlreadyExistingMemberException();
+        }
+        if(form.getBirthDate().isAfter(LocalDate.now().minusYears(16))){
+            throw new UnderAgeException();
+        }
+        shopKeeperAccount = new ShopKeeperAccount(UUID.randomUUID(), form.getName(), form.getMail(), form.getPassword(), form.getBirthDate(), shop);
         shopKeeperAccountRepository.save(shopKeeperAccount, shopKeeperAccount.getId());
         return shopKeeperAccount;
     }
 
     @Override
     public void deleteShopKeeperAccount(ShopKeeperAccount account) {
-        if(shopManager.findShopKeeperAccountById(account.getId()).isPresent()){
+        if(account!=null && shopManager.findShopKeeperAccountById(account.getId()).isPresent()){
             shopRepository.deleteById(account.getId());
         }
+    }
+    @Override
+    public  ShopKeeperAccount findShopkeeperAccountByMail(String mail) {
+        for (ShopKeeperAccount  shopkeeperAccount : shopKeeperAccountRepository.findAll()) {
+            if (shopkeeperAccount.getMail().equals(mail)) {
+                return shopkeeperAccount;
+            }
+        }
+        return null;
     }
 
 }
